% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean}
\usage{
euclidean(x1, x2)
}
\arguments{
\item{x1}{is an integer}

\item{x2}{is an integer}
}
\value{
The function returns the blablabl
}
\description{
Euclidean
}
\examples{
euclidean(100,1000)
}
