% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STATProject.R
\name{print.myr}
\alias{print.myr}
\title{The addition to the generic function print().}
\usage{
\method{print}{myr}(x)
}
\arguments{
\item{x}{function name; The function StatTest is expected to be the only input.}
}
\value{
The output of the test that is readable by the general users.
}
\description{
The addition to the generic function print().
}
\examples{
print.myr(StatTest(dataf="project.csv",test="lm"))
}
